Resolve circular import for decorator
