+++ rUI.py
@@ -2,21 +2,9 @@ from cmd import Cmd
 import include.strings as strings
 from include.device import Device
 import include.testCommand as testCommand
+from include.requireDevice import require_device
 from include.logger import logger
 import os.path, json
-from functools import wraps
-
-# Require Device function decorator. Use this on any method that requires a device to be set
-def require_device(func):
-    @wraps(func)
-    def wrapper(self=None, *arg, **kwargs):
-        # Only run the decorated function if serial is open
-        if self.device.ser.is_open:
-            func(self, *arg, **kwargs)
-        else:
-            print("No device! Please connect to a device with \"connectDevice\"")
-    return wrapper
-
 
 def setLocalOption(data):
     options = {}
